src/requirements.txt
fastapi==0.115.0
pydantic[email]==2.9.2
pydantic-settings==2.5.2
uvicorn[standard]==0.32.0

# Database
sqlalchemy==2.0.36
asyncpg==0.30.0
psycopg2-binary==2.9.10
alembic==1.13.3

# Security & Auth
passlib[bcrypt]==1.7.4
bcrypt==4.2.0
PyJWT==2.9.0
python-multipart==0.0.12
python-jose[cryptography]==3.3.0

# Cloud & Utils
boto3==1.35.0
botocore==1.35.0
python-dotenv==1.0.1

# Testing
pytest==8.3.3
pytest-asyncio==0.24.0
httpx==0.27.2
pytest-mock==3.14.0
aiosqlite==0.20.0